docking state + functions to DockingContainer
 - moved all docking related state and functions to the DockingContainer
 - added state to MoltimateContainer related to properties changed by settings
 - added effects to DockingContainer to make changes to the request form based on props from the MoltimateContainer
