+++ src/docking/DockingContainer.js
@@ -2,6 +2,8 @@ import React, {useState, useEffect} from "react";
 import LigandLibraryContainer from "../ligand_library/LigandLibraryContainer";
 import ImportedLigandsContainer from "../imported_ligands/ImportedLigandsContainer";
 import DockingInfoContainer from "../docking_info/DockingInfoContainer";
+import useForm, {dockRequestURL} from "../util/request"
+import axios from "axios";
 
 import {library_ligands, 
   test_ligands, 
@@ -9,8 +11,11 @@ import {library_ligands,
   fake_docking_data_2} from '../DummyData'
 
 function DockingContainer(props){
-  const {selectedLigands, dockingInProgress, dockingError, setDockingError, dockingResults, setFormValue} = props;
+  const { selectedMacromolecule, dockingCenter, dockingRange} = props;
 
+  //ligands selected for docking. Do not pass this setter (instead use the 
+  //"setSelectedLigands" function)
+  const [selectedLigands, setSelectedLigandsInner] = useState(new Set());
   //autopopulating data on ligands
   const [ libraryLigands, setLibraryLigands ] = useState(library_ligands);
   //data on all the ligands the user has uploaded
@@ -23,6 +28,121 @@ function DockingContainer(props){
   const [dockingConfigs, setDockingConfigs] = useState([]);
   //the docking configuration selected to be viewed
   const [selectedDockingConfig, setSelectedDockingConfig] = useState(null);
+  //true if any ligands are currently being docked
+  const [dockingInProgress, setDockingInProgress] = useState(false);
+  //Each object represnts the results of one docking job
+  //each object contains a ligand, a macromoleculeID, and data
+  const [dockingResults, setDockingResults] = useState(new Array());
+  //Text telling if there is a docking error
+  const [dockingError, setDockingError] = useState(false);
+
+  //this form is used to make docking requests
+  const defaultRequestValues = {
+    center_x: 0,
+    center_y: 0,
+    center_z: 0,
+    size_x: 100,
+    size_y: 100,
+    size_z: 100
+  }
+
+  /**
+   * Polls the backend until it receives the result of a docking job or times out
+   * 
+   * @param {string} requestURL - the url to be polled
+   * @param {int} retryFrequency - number of seconds between retry attempts
+   * @param {int} timeout - number of seconds to retry before giving up
+   * @param {Object} ligand - the ligand being docked
+   * @param {string} macromoleculeID - the ID of the macromolecule involved in docking
+   */
+  function pollDockingResults(requestURL, retryFrequency, timeoutTime, ligand, macromoleculeID){
+
+    console.log(`docking in progress (lvl2)?: ${dockingInProgress}`)
+    console.log(`docking in progress defined? (lvl2)?: ${!(dockingInProgress === undefined)}`)
+      
+    var pollingTimer; 
+
+    function checkDockingResults(){
+      console.log(`docking in progress (lvl3)?: ${dockingInProgress}`)
+      axios.get(requestURL).then( (response) =>{
+        console.log(`Polling Response: ${response.data}`);
+        console.log(`status: ${response.status}`);
+
+        if(response.status == 200){
+          //stop polling
+          clearInterval(pollingTimer);
+
+          //save the results
+          var newDockingResults = new Array().concat(dockingResults);
+          console.log(ligand);
+          console.log(ligand.uniqueID())
+          newDockingResults.push({
+            ligandID: ligand.uniqueID(),
+            macromoleculeID: macromoleculeID,
+            data: response.data
+          });
+          console.log(newDockingResults)
+          setDockingResults(newDockingResults);
+          
+          //indicate docking is complete
+          setDockingInProgress(false);
+          
+        }
+      }).catch((error) => {
+        console.log(`Polling Error: ${error}`);
+        setDockingInProgress(false);
+        clearInterval(pollingTimer)
+        setDockingError("Error Retrieving Docking Data")
+      });
+    }
+
+    pollingTimer = setInterval(checkDockingResults, retryFrequency*1000);
+
+    setTimeout(()=>{clearInterval(pollingTimer)},timeoutTime*1000) 
+  }
+
+  /**
+   * parses and responds to the results from the initial docking request
+   * 
+   * @param {Object} values - the values of the form data in the initial docking request
+   * @param {Object} result - the response from the docking request sent to the backend
+   * @param {Object} result.data - the body of the response from the docking request
+   */
+  function handleDockingResponse(values, result){
+    console.log(`Result Data: ${result.data}`);
+    for(let field in result.data)
+      console.log(`  ${field}: ${result.data[field]}`);
+    console.log(`Result Error: ${result.error}`);
+    if(!result.error){
+      console.log("point 1")
+      if(!values["macromoleculeID"]){
+        console.log("point 2")
+
+        console.error("macromoleculeID was absent from docking request");
+
+        return -1
+      }
+      console.log("point 3")
+      let requestURL = `${dockRequestURL}?jobId=${result.data.jobId}&pdbId=${values.macromoleculeID}`;
+      let retryFrequency = 20;//seconds
+      let timeout = 900;//seconds
+      console.log(`docking in progress (lvl1)?: ${dockingInProgress}`)
+      console.log(`RequestURL: ${requestURL}`)
+
+      //We assume there is only one selected ligand for now - this may be changed at a later time
+      let selectedLigandArray = Array.from(selectedLigands)
+      let selectedLigand = selectedLigandArray[0]
+      console.log(selectedLigand);
+      pollDockingResults(requestURL, retryFrequency, timeout, selectedLigand, values.macromoleculeID)
+    }else{
+      console.log("point 4")
+      console.error(`Error: ${result.error}`);
+    }
+
+    console.log("point 5")
+  }
+
+  const {handleSubmit, setFormValue} = useForm(dockRequestURL,defaultRequestValues,handleDockingResponse);
 
   useEffect(() => {
     
@@ -80,7 +200,33 @@ function DockingContainer(props){
       setSelectedDockingConfig(1);
     }
 
-  },[viewingLigand])
+  },[viewingLigand]);
+
+  useEffect(() => {
+    setFormValue("macromoleculeID",selectedMacromolecule)
+  },[selectedMacromolecule]);
+
+  useEffect(() => {
+    if(dockingCenter){
+      if(dockingCenter.length != 3){
+        console.error(`dockingCenter has length of ${dockingCenter.length}, expected 3`)
+      }
+      setFormValue("center_x",dockingCenter[0])
+      setFormValue("center_y",dockingCenter[1])
+      setFormValue("center_z",dockingCenter[2])
+    }
+  },[dockingCenter]);
+
+  useEffect(() => {
+    if(dockingCenter){
+      if(dockingRange.length != 3){
+        console.error(`dockingRange has length of ${dockingRange.length}, expected 3`)
+      }
+      setFormValue("size_x",dockingRange[0])
+      setFormValue("size_y",dockingRange[1])
+      setFormValue("size_z",dockingRange[2])
+    }
+  },[dockingRange]);
 
   function Ligand(name, structure){
     this.name = name;
@@ -264,6 +410,11 @@ function DockingContainer(props){
     console.log("selected config: " + configSelection[0]);
   }
 
+  function dockLigands(callback){
+    console.log("Docking Request Sent");
+    setDockingInProgress(true);
+  }
+
   /**
    * Used to submit a ligand-protein pair for docking
    * 
@@ -272,7 +423,7 @@ function DockingContainer(props){
    *  number and an (optional) message. 
    */
   function sendDockingRequest(callback){
-    if(!dockingProteinID){
+    if(!selectedMacromolecule){
       console.log("Must select a docking protein");
       return "Must select a protein from the search results to dock a ligand";
     }
